pyproject.toml
[project]
name = "aoc-2023"
version = "0.1"
description = "Solutions for Advent of Code 2023"
readme = "README.md"
requires-python = ">=3.10"
authors = [
  { name = "Emanuel Ferm", email = "emanuel@eferm.com" }
]

[project.optional-dependencies]
cli = [
  "python-dotenv",
  "requests",
]
dev = [
  "ruff",
]

[project.urls]
Homepage = "https://github.com/eferm/aoc-2023"
"Advent of Code Homepage" = "https://adventofcode.com/2023"

[project.scripts]
bootstrap = "cli:bootstrap"

[tool.ruff]
line-length = 79

[tool.ruff.lint]
ignore = [
  "E741",  # Ambiguous variable name: `l`
]